Create benchmark.pwn
This code will run the ini code and create a file with 5 integer fields, 5 float fields, 5 string fields and 5 boolean fields. 

The code will first write the ini file and print time taken in ms.

And after that the file will be readed and time taken to read will be printed.